refactor: Move recommendation tools into folder.
