+++ src/tools/recommendations/get-recommendation-resources.ts
@@ -1,9 +1,9 @@
 import { pathEncode } from "@vantage-sh/vantage-client";
 import z from "zod/v4";
-import { DEFAULT_LIMIT } from "./structure/constants";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
-import paginationData from "./utils/paginationData";
+import { DEFAULT_LIMIT } from "../structure/constants";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
+import paginationData from "../utils/paginationData";
 
 const description = `
 Get a paginated list of all infrastructure resources affected by a specific cost optimization recommendation.
