+++ src/tools/recommendations/index.ts
@@ -1,0 +1,5 @@
+import "./get-recommendation-details";
+import "./get-recommendation-resource-details";
+import "./get-recommendation-resources";
+import "./list-recommendations";
+import "./update-recommendation-view";
