+++ src/tools/recommendations/list-recommendations.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "vitest";
 import tool from "./list-recommendations";
-import { DEFAULT_LIMIT } from "./structure/constants";
+import { DEFAULT_LIMIT } from "../structure/constants";
 import {
   type ExecutionTestTableItem,
   type ExtractOutputSchema,
@@ -9,7 +9,7 @@ import {
   requestsInOrder,
   type SchemaTestTableItem,
   testTool,
-} from "./utils/testing";
+} from "../utils/testing";
 
 type Validators = ExtractValidators<typeof tool>;
 type OutputSchema = ExtractOutputSchema<typeof tool>;
